refac/moving-storages (#124)
* wip

* wip tests

* wip

* test: make tests compile

---------
