+++ pallets/governance/tests/applications.rs
@@ -1,0 +1,1 @@
+
