+++ pallets/subspace/tests/registration.rs
@@ -4,13 +4,11 @@ use std::collections::BTreeSet;
 
 use frame_support::{assert_err, assert_noop, assert_ok};
 use mock::*;
-use pallet_governance_api::GovernanceApi;
 use sp_core::U256;
 
 use log::info;
 use pallet_subspace::{
-    voting::ApplicationStatus, Curator, CuratorApplications, Emission, Error, FloorDelegationFee,
-    GeneralSubnetApplicationCost, MaxAllowedModules, MaxAllowedSubnets, MaxAllowedUids,
+    Emission, Error, FloorDelegationFee, MaxAllowedModules, MaxAllowedSubnets, MaxAllowedUids,
     MaxNameLength, MaxRegistrationsPerBlock, MinNameLength, MinStake, RegistrationsPerBlock, Stake,
     SubnetGaps, SubnetNames, TotalSubnets, N,
 };
@@ -194,60 +192,6 @@ fn register_same_key_twice() {
             register_module(netuid, key, stake),
             Error::<Test>::KeyAlreadyRegistered
         );
-    });
-}
-
-#[test]
-fn test_whitelist() {
-    new_test_ext().execute_with(|| {
-        let key = U256::from(0);
-        let adding_key = U256::from(1);
-        let mut params = SubspaceModule::global_params();
-        params.curator = key;
-        SubspaceModule::set_global_params(params);
-
-        let proposal_cost = GeneralSubnetApplicationCost::<Test>::get();
-        let data = "test".as_bytes().to_vec();
-
-        add_balance(key, proposal_cost + 1);
-        // first submit an application
-        let balance_before = SubspaceModule::get_balance_u64(&key);
-
-        assert_ok!(SubspaceModule::add_dao_application(
-            get_origin(key),
-            adding_key,
-            data.clone(),
-        ));
-
-        let balance_after = SubspaceModule::get_balance_u64(&key);
-        assert_eq!(balance_after, balance_before - proposal_cost);
-
-        // Assert that the proposal is initially in the Pending status
-        for (_, value) in CuratorApplications::<Test>::iter() {
-            assert_eq!(value.status, ApplicationStatus::Pending);
-            assert_eq!(value.user_id, adding_key);
-            assert_eq!(value.data, data);
-        }
-
-        // add key to whitelist
-        assert_ok!(SubspaceModule::add_to_whitelist(
-            get_origin(key),
-            adding_key,
-            1,
-        ));
-
-        let balance_after_accept = SubspaceModule::get_balance_u64(&key);
-
-        assert_eq!(balance_after_accept, balance_before);
-
-        // Assert that the proposal is now in the Accepted status
-        for (_, value) in CuratorApplications::<Test>::iter() {
-            assert_eq!(value.status, ApplicationStatus::Accepted);
-            assert_eq!(value.user_id, adding_key);
-            assert_eq!(value.data, data);
-        }
-
-        assert!(SubspaceModule::is_in_legit_whitelist(&adding_key));
     });
 }
 
@@ -616,58 +560,6 @@ fn test_register_invalid_subnet_name() {
 }
 
 // Subnet 0 Whitelist
-#[test]
-fn test_remove_from_whitelist() {
-    new_test_ext().execute_with(|| {
-        let whitelist_key = U256::from(0);
-        let module_key = U256::from(1);
-        Curator::<Test>::put(whitelist_key);
-
-        let proposal_cost = Test::get_global_governance_configuration().proposal_cost;
-        let data = "test".as_bytes().to_vec();
-
-        // apply
-        add_balance(whitelist_key, proposal_cost + 1);
-        // first submit an application
-        assert_ok!(SubspaceModule::add_dao_application(
-            get_origin(whitelist_key),
-            module_key,
-            data.clone(),
-        ));
-
-        // Add the module_key to the whitelist
-        assert_ok!(SubspaceModule::add_to_whitelist(
-            get_origin(whitelist_key),
-            module_key,
-            1
-        ));
-        assert!(SubspaceModule::is_in_legit_whitelist(&module_key));
-
-        // Remove the module_key from the whitelist
-        assert_ok!(SubspaceModule::remove_from_whitelist(
-            get_origin(whitelist_key),
-            module_key
-        ));
-        assert!(!SubspaceModule::is_in_legit_whitelist(&module_key));
-    });
-}
-
-#[test]
-fn test_invalid_curator() {
-    new_test_ext().execute_with(|| {
-        let whitelist_key = U256::from(0);
-        let invalid_key = U256::from(1);
-        let module_key = U256::from(2);
-        Curator::<Test>::put(whitelist_key);
-
-        // Try to add to whitelist with an invalid curator key
-        assert_noop!(
-            SubspaceModule::add_to_whitelist(get_origin(invalid_key), module_key, 1),
-            Error::<Test>::NotCurator
-        );
-        assert!(!SubspaceModule::is_in_legit_whitelist(&module_key));
-    });
-}
 
 #[test]
 fn new_subnet_reutilized_removed_netuid_if_total_is_bigger_than_removed() {
